refactor: split `Dialect` from `DictWordMetadata`
