+++ harper-asciidoc/tests/asciidoc_tests.rs
@@ -1,7 +1,8 @@
 use harper_asciidoc::AsciidocParser;
+use harper_core::Document;
+use harper_core::dialect::Dialect;
 use harper_core::linting::{LintGroup, Linter};
 use harper_core::spell::FstDictionary;
-use harper_core::{Dialect, Document};
 
 /// Creates a unit test checking Asciidoc source code parsing.
 macro_rules! create_test {
