+++ harper-comments/tests/language_support.rs
@@ -1,10 +1,11 @@
 use std::path::Path;
 
 use harper_comments::CommentParser;
+use harper_core::Document;
+use harper_core::dialect::Dialect;
 use harper_core::linting::{LintGroup, Linter};
 use harper_core::parsers::MarkdownOptions;
 use harper_core::spell::FstDictionary;
-use harper_core::{Dialect, Document};
 
 /// Creates a unit test checking that the linting of a source file in
 /// `language_support_sources` produces the expected number of lints.
