+++ harper-core/src/linting/analog_acoustic_bike.rs
@@ -1,5 +1,6 @@
 use crate::{
-    Dialect, Lint, Token, TokenStringExt,
+    Lint, Token, TokenStringExt,
+    dialect::Dialect,
     expr::{Expr, SequenceExpr},
     linting::{ExprLinter, LintKind, Suggestion, expr_linter::Chunk},
     patterns::IndefiniteArticle,
@@ -72,7 +73,7 @@ impl ExprLinter for AnalogAcousticBike {
 
 #[cfg(test)]
 mod tests {
-    use crate::{Dialect, linting::tests::assert_good_and_bad_suggestions};
+    use crate::{dialect::Dialect, linting::tests::assert_good_and_bad_suggestions};
 
     use super::AnalogAcousticBike;
