+++ harper-core/src/linting/lint_group/flat_config.rs
@@ -7,7 +7,7 @@ use hashbrown::HashMap;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 
 use super::LintGroup;
-use crate::Dialect;
+use crate::dialect::Dialect;
 use crate::spell::MutableDictionary;
 
 fn ser_ordered<S>(map: &HashMap<String, Option<bool>>, ser: S) -> Result<S::Ok, S::Error>
