+++ harper-desktop/src-tauri/src/config/mod.rs
@@ -3,7 +3,8 @@ mod integration;
 
 pub use error::Error;
 use harper_core::{
-    Dialect, IgnoredLints,
+    IgnoredLints,
+    dialect::Dialect,
     linting::{FlatConfig, LintGroup},
     spell::{FstDictionary, MergedDictionary, MutableDictionary},
 };
