+++ harper-jjdescription/tests/run_tests.rs
@@ -1,7 +1,8 @@
+use harper_core::Document;
+use harper_core::dialect::Dialect;
 use harper_core::linting::{LintGroup, Linter};
 use harper_core::parsers::MarkdownOptions;
 use harper_core::spell::FstDictionary;
-use harper_core::{Dialect, Document};
 use harper_jjdescription::JJDescriptionParser;
 
 /// Creates a unit test checking that the linting of a git commit document (in
