+++ harper-wasm/src/lib.rs
@@ -95,14 +95,14 @@ pub enum Dialect {
     Indian,
 }
 
-impl From<Dialect> for harper_core::Dialect {
+impl From<Dialect> for harper_core::dialect::Dialect {
     fn from(dialect: Dialect) -> Self {
         match dialect {
-            Dialect::American => harper_core::Dialect::American,
-            Dialect::Canadian => harper_core::Dialect::Canadian,
-            Dialect::Australian => harper_core::Dialect::Australian,
-            Dialect::British => harper_core::Dialect::British,
-            Dialect::Indian => harper_core::Dialect::Indian,
+            Dialect::American => harper_core::dialect::Dialect::American,
+            Dialect::Canadian => harper_core::dialect::Dialect::Canadian,
+            Dialect::Australian => harper_core::dialect::Dialect::Australian,
+            Dialect::British => harper_core::dialect::Dialect::British,
+            Dialect::Indian => harper_core::dialect::Dialect::Indian,
         }
     }
 }
